requirements.txt
fastapi>=0.61.1
uvicorn>=0.12.1
jinja2>=2.11.2
aiofiles>=0.5.0
beautifulsoup4>=4.9.3
python-multipart>=0.0.5
numpy>=1.19.4
tweepy>=3.10.0
SQLAlchemy>=1.4.9
psycopg2>=2.8.6
requests
Pillow
line-bot-sdk
schedule
